wip: simplify outbound operation validation
