+++ .agents/scripts/knowledge_social_operations.py
@@ -6,7 +6,6 @@
 from __future__ import annotations
 
 import argparse
-import hashlib
 import json
 import os
 import sqlite3
@@ -22,9 +21,14 @@ from _knowledge_social_notifications import (
     project_notifications,
     set_notification_status,
 )
+from _knowledge_social_operation_files import (
+    OperationFileError,
+    private_media_digest,
+    read_private_body,
+    read_private_subject,
+)
 from _knowledge_social_outbound import (
     ACTIONS,
-    MAX_PAYLOAD_BYTES,
     MAX_SUBJECT_BYTES,
     ClaimedOperation,
     OperationIntent,
@@ -54,7 +58,7 @@ from _knowledge_social_outbound_runtime import (
     record_provider_checkpoint,
 )
 from knowledge_corpus_catalog import DEFAULT_ALIAS, authorized_scope
-from knowledge_corpus_context import CatalogError, validate_private_file
+from knowledge_corpus_context import CatalogError
 from knowledge_social_store import (
     SocialStoreError,
     connect,
@@ -84,79 +88,6 @@ def _clock(args: argparse.Namespace) -> int:
             raise OperationsError("test clock must be a non-negative epoch")
         return override
     return int(time.time())
-
-
-def _open_private_body(path: Path) -> tuple[int, os.stat_result]:
-    try:
-        validate_private_file(path, "outbound body", repair=False)
-        before = path.lstat()
-        flags = os.O_RDONLY
-        if hasattr(os, "O_NOFOLLOW"):
-            flags |= os.O_NOFOLLOW
-        descriptor = os.open(path, flags)
-    except (CatalogError, OSError) as error:
-        raise OperationsError("outbound body is unavailable or unsafe") from error
-    return descriptor, before
-
-
-def _read_private_bytes(descriptor: int, before: os.stat_result) -> bytes:
-    try:
-        after = os.fstat(descriptor)
-        if (before.st_dev, before.st_ino) != (after.st_dev, after.st_ino):
-            raise OperationsError("outbound body replacement detected")
-        with os.fdopen(descriptor, "rb") as handle:
-            descriptor = -1
-            payload = handle.read(MAX_PAYLOAD_BYTES + 1)
-    except OSError as error:
-        raise OperationsError("outbound body is unavailable or unsafe") from error
-    finally:
-        if descriptor >= 0:
-            os.close(descriptor)
-    return payload
-
-
-def _read_private_body(path: Path) -> str:
-    descriptor, before = _open_private_body(path)
-    payload = _read_private_bytes(descriptor, before)
-    if len(payload) > MAX_PAYLOAD_BYTES:
-        raise OperationsError("outbound body exceeds the private payload limit")
-    try:
-        return payload.decode("utf-8")
-    except UnicodeDecodeError as error:
-        raise OperationsError("outbound body must be UTF-8") from error
-
-
-def _read_private_subject(path: Path) -> str:
-    subject = _read_private_body(path)
-    if subject.endswith("\n"):
-        subject = subject[:-1]
-        if subject.endswith("\r"):
-            subject = subject[:-1]
-    return subject
-
-
-def _private_media_digest(path: Path) -> tuple[str, str]:
-    """Return an absolute private media path and a bounded streaming digest."""
-    try:
-        validate_private_file(path, "outbound media", repair=False)
-        resolved = path.resolve(strict=True)
-        before = resolved.stat()
-        if not resolved.is_file() or before.st_size < 1 or before.st_size > 2 * 1024**3:
-            raise OperationsError("outbound media is unavailable or outside the size limit")
-        digest = hashlib.sha256()
-        with resolved.open("rb") as handle:
-            while chunk := handle.read(1024 * 1024):
-                digest.update(chunk)
-        after = resolved.stat()
-        if (before.st_dev, before.st_ino, before.st_size) != (
-            after.st_dev,
-            after.st_ino,
-            after.st_size,
-        ):
-            raise OperationsError("outbound media replacement detected")
-    except (CatalogError, OSError) as error:
-        raise OperationsError("outbound media is unavailable or unsafe") from error
-    return str(resolved), digest.hexdigest()
 
 
 def _managed_context(args: argparse.Namespace) -> tuple[str, Path]:
@@ -323,10 +254,10 @@ def _handle_operation_create(
     current_time: int | None,
 ) -> dict[str, Any]:
     created_at = _required_now(current_time)
-    payload = _read_private_body(args.body_file) if args.body_file else None
-    subject = _read_private_subject(args.subject_file) if args.subject_file else None
+    payload = read_private_body(args.body_file) if args.body_file else None
+    subject = read_private_subject(args.subject_file) if args.subject_file else None
     media_path, media_sha256 = (
-        _private_media_digest(args.media_file) if args.media_file else (None, None)
+        private_media_digest(args.media_file) if args.media_file else (None, None)
     )
     if subject is not None and len(subject.encode("utf-8")) > MAX_SUBJECT_BYTES:
         raise OperationsError("outbound subject exceeds the private subject limit")
@@ -637,6 +568,7 @@ def main() -> int:
     except (
         CatalogError,
         OSError,
+        OperationFileError,
         OperationsError,
         ProviderAdapterError,
         ProviderIdentityError,
